Commit change for heroku. Replace double to float

+++ common/src/main/java/com/example/common/CityTemp.kt
@@ -2,4 +2,4 @@ package com.example.common
 
 import java.io.Serializable
 
-data class CityTemp (val city : String, val temp : Double = -1.0, val id : Long = -1) : Serializable
+data class CityTemp (val city : String, val temp : Float = -1.0F, val id : Long = -1) : Serializable
